LPS-56048 Move MVC actions into module
