+++ modules/apps/message-boards/message-boards-web/src/com/liferay/message/boards/web/portlet/action/BanUserMVCActionCommand.java
@@ -12,69 +12,76 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.action;
+package com.liferay.message.boards.web.portlet.action;
 
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
 import com.liferay.portal.kernel.servlet.SessionErrors;
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
+import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.security.auth.PrincipalException;
-import com.liferay.portal.util.PortalUtil;
+import com.liferay.portal.service.ServiceContext;
+import com.liferay.portal.service.ServiceContextFactory;
 import com.liferay.portal.util.PortletKeys;
-import com.liferay.portlet.messageboards.service.MBCategoryServiceUtil;
+import com.liferay.portlet.messageboards.model.MBBan;
+import com.liferay.portlet.messageboards.service.MBBanServiceUtil;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
 
 /**
- * @author Eduardo Garcia
+ * @author Michael Young
  */
 @OSGiBeanProperties(
 	property = {
 		"javax.portlet.name=" + PortletKeys.MESSAGE_BOARDS,
 		"javax.portlet.name=" + PortletKeys.MESSAGE_BOARDS_ADMIN,
-		"mvc.command.name=/message_boards/move_category"
+		"mvc.command.name=/message_boards/ban_user"
 	},
 	service = MVCActionCommand.class
 )
-public class MoveCategoryMVCActionCommand extends BaseMVCActionCommand {
+public class BanUserMVCActionCommand extends BaseMVCActionCommand {
+
+	protected void banUser(ActionRequest actionRequest) throws Exception {
+		long banUserId = ParamUtil.getLong(actionRequest, "banUserId");
+
+		ServiceContext serviceContext = ServiceContextFactory.getInstance(
+			MBBan.class.getName(), actionRequest);
+
+		MBBanServiceUtil.addBan(banUserId, serviceContext);
+	}
 
 	@Override
 	protected void doProcessAction(
 			ActionRequest actionRequest, ActionResponse actionResponse)
 		throws Exception {
 
+		String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
+
 		try {
-			moveCategory(actionRequest, actionResponse);
-
-			String redirect = PortalUtil.escapeRedirect(
-				ParamUtil.getString(actionRequest, "redirect"));
-
-			if (Validator.isNotNull(redirect)) {
-				actionResponse.sendRedirect(redirect);
+			if (cmd.equals("ban")) {
+				banUser(actionRequest);
+			}
+			else if (cmd.equals("unban")) {
+				unbanUser(actionRequest);
 			}
 		}
 		catch (PrincipalException pe) {
 			SessionErrors.add(actionRequest, pe.getClass());
+
+			actionResponse.setRenderParameter(
+				"mvcPath", "/html/portlet/message_boards/error.jsp");
 		}
 	}
 
-	protected void moveCategory(
-			ActionRequest actionRequest, ActionResponse actionResponse)
-		throws Exception {
+	protected void unbanUser(ActionRequest actionRequest) throws Exception {
+		long banUserId = ParamUtil.getLong(actionRequest, "banUserId");
 
-		long categoryId = ParamUtil.getLong(actionRequest, "mbCategoryId");
+		ServiceContext serviceContext = ServiceContextFactory.getInstance(
+			MBBan.class.getName(), actionRequest);
 
-		long parentCategoryId = ParamUtil.getLong(
-			actionRequest, "parentCategoryId");
-
-		boolean mergeWithParentCategory = ParamUtil.getBoolean(
-			actionRequest, "mergeWithParentCategory");
-
-		MBCategoryServiceUtil.moveCategory(
-			categoryId, parentCategoryId, mergeWithParentCategory);
+		MBBanServiceUtil.deleteBan(banUserId, serviceContext);
 	}
 
 }
