+++ modules/apps/message-boards/message-boards-web/src/com/liferay/message/boards/web/portlet/action/EditCategoryMVCActionCommand.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.action;
+package com.liferay.message.boards.web.portlet.action;
 
 import com.liferay.portal.kernel.captcha.CaptchaConfigurationException;
 import com.liferay.portal.kernel.captcha.CaptchaMaxChallengesException;
