+++ modules/apps/message-boards/message-boards-web/src/com/liferay/message/boards/web/portlet/action/FindCategoryAction.java
@@ -12,30 +12,41 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.action;
+package com.liferay.message.boards.web.portlet.action;
 
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
+import com.liferay.portal.kernel.struts.BaseStrutsAction;
+import com.liferay.portal.kernel.struts.StrutsAction;
+import com.liferay.portal.struts.FindActionHelper;
 import com.liferay.portal.util.PortletKeys;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
 /**
- * @author Adolfo Pérez
+ * @author Brian Wing Shun Chan
  */
 @OSGiBeanProperties(
 	property = {
 		"javax.portlet.name=" + PortletKeys.MESSAGE_BOARDS,
 		"javax.portlet.name=" + PortletKeys.MESSAGE_BOARDS_ADMIN,
-		"mvc.command.name=/message_boards/view_deleted_message_attachments"
+		"path=/message_boards/find_category"
 	},
-	service = MVCRenderCommand.class
+	service = StrutsAction.class
 )
-public class EditMessageAttachmentsMVCRenderCommand
-	extends GetMessageMVCRenderCommand {
+public class FindCategoryAction extends BaseStrutsAction {
 
 	@Override
-	protected String getPath() {
-		return
-			"/html/portlet/message_boards/view_deleted_message_attachments.jsp";
+	public String execute(
+			HttpServletRequest request, HttpServletResponse response)
+		throws Exception {
+
+		_findActionHelper.execute(request, response);
+
+		return null;
 	}
+
+	private final FindActionHelper _findActionHelper =
+		new CategoryFindActionHelper();
 
 }
