+++ modules/apps/message-boards/message-boards-web/src/com/liferay/message/boards/web/portlet/action/RSSAction.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.action;
+package com.liferay.message.boards.web.portlet.action;
 
 import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
