+++ modules/apps/message-boards/message-boards-web/src/com/liferay/message/boards/web/portlet/action/SearchMVCRenderCommand.java
@@ -12,41 +12,32 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.action;
+package com.liferay.message.boards.web.portlet.action;
 
+import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
-import com.liferay.portal.kernel.struts.BaseStrutsAction;
-import com.liferay.portal.kernel.struts.StrutsAction;
-import com.liferay.portal.struts.FindActionHelper;
 import com.liferay.portal.util.PortletKeys;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import javax.portlet.RenderRequest;
+import javax.portlet.RenderResponse;
 
 /**
- * @author Brian Wing Shun Chan
+ * @author Adolfo Pérez
  */
 @OSGiBeanProperties(
 	property = {
 		"javax.portlet.name=" + PortletKeys.MESSAGE_BOARDS,
 		"javax.portlet.name=" + PortletKeys.MESSAGE_BOARDS_ADMIN,
-		"path=/message_boards/find_thread"
-	},
-	service = StrutsAction.class
+		"mvc.command.name=/message_boards/search"
+	}
 )
-public class FindThreadAction extends BaseStrutsAction {
+public class SearchMVCRenderCommand implements MVCRenderCommand {
 
 	@Override
-	public String execute(
-			HttpServletRequest request, HttpServletResponse response)
-		throws Exception {
+	public String render(
+		RenderRequest renderRequest, RenderResponse renderResponse) {
 
-		_findActionHelper.execute(request, response);
-
-		return null;
+		return "/portlet/message_boards/search.jsp";
 	}
-
-	private final FindActionHelper _findActionHelper =
-		new ThreadFindActionHelper();
 
 }
