+++ modules/apps/message-boards/message-boards-web/src/com/liferay/message/boards/web/portlet/action/ThreadFindActionHelper.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.action;
+package com.liferay.message.boards.web.portlet.action;
 
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.struts.BaseFindActionHelper;
