+++ modules/apps/message-boards/message-boards-web/src/com/liferay/message/boards/web/portlet/action/ViewMVCRenderCommand.java
@@ -12,31 +12,32 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.action;
+package com.liferay.message.boards.web.portlet.action;
 
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
 import com.liferay.portal.kernel.servlet.SessionErrors;
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
 import com.liferay.portal.security.auth.PrincipalException;
 import com.liferay.portal.util.PortletKeys;
-import com.liferay.portlet.messageboards.NoSuchMessageException;
+import com.liferay.portlet.messageboards.BannedUserException;
+import com.liferay.portlet.messageboards.NoSuchCategoryException;
 
 import javax.portlet.PortletException;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
 
 /**
- * @author Adolfo Pérez
+ * @author Brian Wing Shun Chan
  */
 @OSGiBeanProperties(
 	property = {
 		"javax.portlet.name=" + PortletKeys.MESSAGE_BOARDS,
 		"javax.portlet.name=" + PortletKeys.MESSAGE_BOARDS_ADMIN,
-		"mvc.command.name=/message_boards/move_thread"
+		"mvc.command.name=/", "mvc.command.name=/message_boards/view"
 	},
 	service = MVCRenderCommand.class
 )
-public class MoveThreadMVCRenderCommand implements MVCRenderCommand {
+public class ViewMVCRenderCommand implements MVCRenderCommand {
 
 	@Override
 	public String render(
@@ -44,9 +45,11 @@ public class MoveThreadMVCRenderCommand implements MVCRenderCommand {
 		throws PortletException {
 
 		try {
-			ActionUtil.getThreadMessage(renderRequest);
+			ActionUtil.getCategory(renderRequest);
 		}
-		catch (NoSuchMessageException | PrincipalException e) {
+		catch (BannedUserException | NoSuchCategoryException |
+				PrincipalException e) {
+
 			SessionErrors.add(renderRequest, e.getClass());
 
 			return "/html/portlet/message_boards/error.jsp";
@@ -58,7 +61,7 @@ public class MoveThreadMVCRenderCommand implements MVCRenderCommand {
 			throw new PortletException(e);
 		}
 
-		return "/html/portlet/message_boards/move_thread.jsp";
+		return "/html/portlet/message_boards/view.jsp";
 	}
 
 }
