+++ modules/apps/message-boards/message-boards-web/src/com/liferay/message/boards/web/portlet/action/ViewMessageMVCRenderCommand.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.action;
+package com.liferay.message.boards.web.portlet.action;
 
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
 import com.liferay.portal.kernel.servlet.SessionErrors;
