+++ portal-impl/src/portal.properties
@@ -834,7 +834,6 @@
         META-INF/cluster-spring.xml,\
         META-INF/comment-spring.xml,\
         META-INF/form-navigator-spring.xml,\
-        META-INF/message-boards-spring.xml,\
         META-INF/model-listener-spring.xml,\
         META-INF/notifications-spring.xml,\
         META-INF/poller-spring.xml,\
